# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/events": {
      "get": {
        "operationId": "EventsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "EventsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
...
    "/api/payments/webhook": {
      "post": {
        "operationId": "PaymentsController_webhook",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/payments/{id}": {
      "get": {
        "operationId": "PaymentsController_getPayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/events/{eventId}/guest-lists": {
      "post": {
        "operationId": "EventGuestListsController_create",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGuestListDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      },
      "get": {
        "operationId": "EventGuestListsController_list",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/guest-lists/{id}/entries": {
      "post": {
        "operationId": "GuestListsController_addEntry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddEntryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/guest-lists/{listId}/entries/{entryId}/pass": {
      "post": {
        "operationId": "GuestListsController_issuePass",
        "parameters": [
          {
            "name": "listId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "omni-dance API",
    "description": "API de la plataforma omni-dance (SBK Santiago): eventos, QR, sesiones, tickets, social, academias, gamificación, RBAC y admin.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "omnidance_session"
      }
    },
    "schemas": {
...
      "CreateEventDto": {
        "type": "object",
        "properties": {}
      },
...
      "WebhookDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateGuestListDto": {
        "type": "object",
        "properties": {}
      },
      "AddEntryDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}